requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.2
